Split main-step coverage probes into coverage-probe.js; name CoverageContext typedef
